refactor(polly): durable training capture + sibling-bundle LLM helper
Two follow-ups to the initial commerce bridge:

1. Cross-function require risk eliminated. Previously
   api/polly/chat.js did `require("../agent/chat")` to reuse the
   LLM router. @vercel/node bundles each function entrypoint
   independently with ncc; cross-function requires aren't
   guaranteed to be hoisted into the polly bundle and would 500
   at runtime with "Cannot find module". Extracted the router to
   api/_chat_llm.js (sibling helper, same pattern as
   _agent_common.js) and rewired both api/agent/chat.js and
   api/polly/chat.js to require from there. Public surface of
   api/agent/chat.js is unchanged: the `_private` export still
   exposes chatConfig/routeChat/etc.

2. Training capture is now durable. Stdout-only logging gets lost
   when Vercel rolls function logs. Added api/_polly_train_capture.js
   which fires a GitHub repository_dispatch event of type
   `polly_training_turn` whenever a consented turn comes in. The
   new `.github/workflows/polly-training-capture.yml` listens for
   that event and appends the validated JSON line to
   training-data/polly-chat-live/{YYYY-MM}.jsonl on main, then
   commits. Path is non-blocking — never throws into the chat
   path; falls back to stdout-only when GITHUB_TOKEN/POLLY_TRAIN_GITHUB_TOKEN
   isn't set or POLLY_TRAIN_DISPATCH_ENABLED=false.

Tests: 23/23 (added 3 cases for the dispatcher's no-token, disabled,
and event-type contracts).
